% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{add_abstractsToCorpus}
\alias{add_abstractsToCorpus}
\title{Add retrieved abstract to corpus}
\usage{
add_abstractsToCorpus(in_corpus, scopusAbstracts, wosAbstracts)
}
\arguments{
\item{in_corpus}{corpus database with ArticleURL column}

\item{scopusAbstracts}{abstracts extracted from Scopus database}

\item{wosAbstracts}{abstracts extracted from Web of Science database}
}
\value{
corpus corpus database with a new column abstract
}
\description{
Add retrieved abstract to corpus
}
