% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{align.englishCorpus}
\alias{align.englishCorpus}
\title{Align englishCorpus with matching records in both databases and assign webscrapped abstracts}
\usage{
align.englishCorpus(englishCorpus, EndNoteIdLDA, EndNoteIdcorpus, in_corpus)
}
\arguments{
\item{englishCorpus}{corpus of document}

\item{EndNoteIdLDA}{document id in the corpus from the query}

\item{EndNoteIdcorpus}{document id in the topic model}
}
\value{
englishCorpus with matching records in both databases and webscrapped abstracts
}
\description{
Align englishCorpus with matching records in both databases and assign webscrapped abstracts
}
