% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{align.humanReadingTopicModel}
\alias{align.humanReadingTopicModel}
\title{identifies the subset of paper with validation data and align databases}
\usage{
align.humanReadingTopicModel(titleInd, validationHumanReading, topicDocs, DTM)
}
\arguments{
\item{titleInd}{cross-walked indices between human-reading database and topic model}

\item{validationHumanReading}{human-reading database}

\item{topicDocs}{results from the topic model}

\item{DTM}{document-term matrix derived from topic modelled corpus}
}
\value{
list with four elements: titleInd, validationHumanReading, validationTopicDocs, validationDTM
}
\description{
identifies the subset of paper with validation data and align databases
}
