% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{align_dataWithEndNoteIdcorpus}
\alias{align_dataWithEndNoteIdcorpus}
\title{Align databases based on shared ID}
\usage{
align_dataWithEndNoteIdcorpus(data, EndNoteIdcorpus, EndNoteIdLDA)
}
\arguments{
\item{EndNoteIdcorpus}{document id in the topic model}

\item{EndNoteIdLDA}{document id in the corpus from the query}
}
\value{
EndNoteIdcorpus with matching records in both databases
}
\description{
Align databases based on shared ID
}
