% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_formatting.R
\name{article_selection}
\alias{article_selection}
\title{Randomly select articles for human-reading}
\usage{
article_selection(full.names, non_duplicate_index, ratio = 0.2, .seed = 89)
}
\arguments{
\item{full.names}{file.paths to corpus document}

\item{non_duplicate_index}{index of non-duplicated documents}

\item{ratio}{ratio of total number of documents to select; defaults to 20\%}

\item{.seed}{random seed for reproducible results; default to 89}
}
\description{
Randomly select articles for human-reading
}
