% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_download.R
\name{assign_articles_to_players}
\alias{assign_articles_to_players}
\title{Create a \code{.csv} files with the information to download the documents}
\usage{
assign_articles_to_players(language, number_of_players = 8)
}
\arguments{
\item{language}{character, one of "english", "spanish", "portuguese"}

\item{number_of_players}{numeric, number of downloarders}
}
\description{
Create a \code{.csv} files with the information to download the documents
}
