% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_formatting.R
\name{assign_articles_to_readers}
\alias{assign_articles_to_readers}
\title{Chunk the selected documents between a given number of human readers and copy them into a given directory}
\usage{
assign_articles_to_readers(select_ind, number_of_readers, out_dir)
}
\arguments{
\item{select_ind}{index of the selected documents}

\item{number_of_readers}{numeric, number of readers}

\item{out_dir}{directory to output the files}
}
\description{
Chunk the selected documents between a given number of human readers and copy them into a given directory
}
