% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{check_duplicate_row}
\alias{check_duplicate_row}
\title{Return the unique rows of a data.frame.}
\usage{
check_duplicate_row(citation_dataframe)
}
\arguments{
\item{citation_dataframe}{a data.frame}
}
\value{
a data.frame with non-duplicated rows
}
\description{
Return the unique rows of a data.frame.
}
