% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{check_duplicate_title}
\alias{check_duplicate_title}
\title{Return a data.frame with unique Titles.}
\usage{
check_duplicate_title(citation_dataframe)
}
\arguments{
\item{citation_dataframe}{a data.frame with a \code{Title} column}
}
\value{
a data.frame with non-duplicated \code{Title}
}
\description{
Return a data.frame with unique Titles.
}
