% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clStab}
\alias{clStab}
\title{Evaluates the cluster stability}
\usage{
clStab(df, method = "kmeans", kmax = 10)
}
\arguments{
\item{df}{a data.frame}

\item{method}{character, one of "kmeans", "hierarchical"}

\item{kmax}{numeric, the maximum number of cluster to test}
}
\value{
a named list with two elements: \code{stab} the result from the stability function \code{clValid::clValid} and \code{optim_df} a data.frame with optimum values following the one-standard-deviation rule
}
\description{
Evaluates the cluster stability
}
