% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{consolidate_LDA_results}
\alias{consolidate_LDA_results}
\title{Consolidates lda results by adding year and country prediction to the topicDocs}
\usage{
consolidate_LDA_results(
  theme_type = "theme",
  description = NULL,
  save = FALSE,
  .topicDocs = topicDocs,
  .in_corpus = in_corpus,
  .predCountry = predCountry
)
}
\arguments{
\item{theme_type}{one of topic_name, theme, NSF_specific, NSF_general, passed to make_df_docs}

\item{description}{one of spatial_scale, methods, water budget, passed to make_df_docs}

\item{save}{boolean, if true, write the results, if false, return the consolidated results}

\item{topicDocs}{the output from the topicdata, aligned with the corpus}

\item{in_corpus}{the corpus database containing the variable Year}

\item{predCountry}{the complete probabilities predicted from ML}
}
\value{
see save
}
\description{
Consolidates lda results by adding year and country prediction to the topicDocs
}
