% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{count_nas}
\alias{count_nas}
\title{Count the number of \code{NA}}
\usage{
count_nas(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
numeric, number of \code{NA} found
}
\description{
Count the number of \code{NA}
}
