% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_country}
\alias{diversity_country}
\title{Calculates the diversity}
\usage{
diversity_country(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame
}
\description{
Calculates the diversity
}
