% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_LAC}
\alias{diversity_LAC}
\title{Calculates a diversity over the entire LAC region}
\usage{
diversity_LAC(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame
}
\description{
Calculates a diversity over the entire LAC region
}
