% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{EDA_trainingData}
\alias{EDA_trainingData}
\title{Performs a simple visualization of multilabel training data using mldr package}
\usage{
EDA_trainingData(
  trainingData,
  validationHumanReadingDTM,
  humanReadingTrainingLabels
)
}
\arguments{
\item{trainingData}{data.frame of training data}

\item{validationHumanReadingDTM}{document-term matrix from human reading}

\item{humanReadingTrainingLabels}{labels from human-reading}
}
\value{
a data.frame of comparison between the human reading and the webscrapped labels
}
\description{
Performs a simple visualization of multilabel training data using mldr package
}
