% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{filter_by_country}
\alias{filter_by_country}
\title{Filter rows of a data.frame by country}
\usage{
filter_by_country(
  .df = df,
  country = "Brazil",
  country_type = "Country_current"
)
}
\arguments{
\item{.df}{a data.frame}

\item{country}{character, country to search for}

\item{country_type}{character, name of the column in which to search for \code{country}}
}
\value{
a data.frame filtered by rows containing \code{country}
}
\description{
Filter rows of a data.frame by country
}
