% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{filter_columns}
\alias{filter_columns}
\title{Filter columns of a data.frame}
\usage{
filter_columns(df, colname_list = c("Country_current", "Years"))
}
\arguments{
\item{df}{a data.frame}

\item{colname_list}{character, the names of the column(s) to filter}
}
\value{
a data.frame with the column(s) present in \code{colname_list}
}
\description{
Filter columns of a data.frame
}
