% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{filter_dfm}
\alias{filter_dfm}
\title{Filter the complete document-feature matrix to retain all features with occurence higher than the lowest occurence of country tokens.
This function mainly serves to limit the size of the document-feature matrix}
\usage{
filter_dfm(obj_dfm, country_tokens = NULL)
}
\arguments{
\item{obj_dfm}{a document-feature matrix}

\item{country_tokens}{the tokens with which the document-feature matrix is filtered}
}
\value{
a filtered document-feature matrix
}
\description{
Filter the complete document-feature matrix to retain all features with occurence higher than the lowest occurence of country tokens.
This function mainly serves to limit the size of the document-feature matrix
}
