% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_formatting.R
\name{fix_names}
\alias{fix_names}
\title{Fix the format of the document names from an existing database}
\usage{
fix_names(lang, full_names, language_df = NULL)
}
\arguments{
\item{lang}{character defining the language, one of: "english", "portuguese", "spanish"}

\item{full_names}{character, the list of full file.path to the document files}

\item{language_df}{a named list of data.frame, at least one name should correspond to \code{lang}. The resulting data.frame should have a \code{pdfs} column}
}
\value{
a named list with two elements: \code{names} and \code{full_names}
}
\description{
Fix the format of the document names from an existing database
}
