% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{format_data4get_ps}
\alias{format_data4get_ps}
\title{Format the data to the format expected by get_ps}
\usage{
format_data4get_ps(data)
}
\arguments{
\item{data}{data to be formatted}
}
\value{
list with two named elements: data and labels
}
\description{
Format the data to the format expected by get_ps
}
