% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_QA.R
\name{generate_label_df}
\alias{generate_label_df}
\title{General a data.frame of labels for a violin plot}
\usage{
generate_label_df(HSD, flev, d, y, fact = 1, yadjust = 0)
}
\arguments{
\item{HSD}{results from  \code{TukeyHSD()}}

\item{flev}{\code{factor} to use for the Tukey levels}

\item{d}{data.frame}

\item{y}{second \code{factor} for the Tukey levels}

\item{fact}{multiplies the maximum value to adjust graphical position}

\item{yadjust}{absolute adjustment position}
}
\value{
data.frame
}
\description{
General a data.frame of labels for a violin plot
}
