% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get.allMetadata}
\alias{get.allMetadata}
\title{Extract metadata from Scopus or Web of Science identifiers}
\usage{
get.allMetadata(
  documentID,
  fun = get.scopusAbstract,
  newpass = FALSE,
  metadata = NULL,
  ind_null = NULL
)
}
\arguments{
\item{documentID}{list of URLs for Scopus database or DOIs for Web of Science database}

\item{fun}{function, either get.scopusAbstract, get.wosAbstract or get.wosFullResult}

\item{newpass}{boolean, default to FALSE, if TRUE, ind_null and metadata are expected to be set}

\item{metadata}{list, list of already retrieved metadata}

\item{ind_null}{list, flags for failure to retrieve}
}
\value{
a list of metadata: only abstracts or complete metadata
}
\description{
Extract metadata from Scopus or Web of Science identifiers
}
