% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get.boolean_AuthKeywords}
\alias{get.boolean_AuthKeywords}
\title{Transform the author keywords into a multilabel dataset}
\usage{
get.boolean_AuthKeywords(wosAuthKeywords, relevant_countries)
}
\arguments{
\item{wosAuthKeywords}{author keywords}

\item{relevant_countries}{labels to search for}
}
\value{
a data.frame
}
\description{
Transform the author keywords into a multilabel dataset
}
