% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get.chainingOrder}
\alias{get.chainingOrder}
\title{Get chaining order from MLDR}
\usage{
get.chainingOrder(
  trainingData,
  validationHumanReadingDTM,
  scale_type = "location",
  aggregated_labels = FALSE
)
}
\arguments{
\item{trainingData}{data.frame of training data}

\item{validationHumanReadingDTM}{document-term matrix from human reading}

\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}

\item{aggregated_labels}{logical}
}
\description{
Get chaining order from MLDR
}
