% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get.EndNoteIdLDA}
\alias{get.EndNoteIdLDA}
\title{get document ID from LDA corpus database}
\usage{
get.EndNoteIdLDA(englishCorpus)
}
\arguments{
\item{englishCorpus}{LDA corpus database}
}
\value{
list of document IDs
}
\description{
get document ID from LDA corpus database
}
