% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get.relevantCountries}
\alias{get.relevantCountries}
\title{Extract countries names to be searched for in the author keywords}
\usage{
get.relevantCountries()
}
\value{
List of country names
}
\description{
Extract countries names to be searched for in the author keywords
}
