% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get.tuningPar}
\alias{get.tuningPar}
\title{Retrieve the most frequent best tuned hyper-parameters
Ties are broken by taking the mininum}
\usage{
get.tuningPar(bmr, tuning_par)
}
\arguments{
\item{bmr}{a MLR benchmark}

\item{tuning_par}{a tuning parameter name}
}
\description{
Retrieve the most frequent best tuned hyper-parameters
Ties are broken by taking the mininum
}
