% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get.wosFullResult}
\alias{get.wosFullResult}
\title{Use DOI to extract metadata}
\usage{
get.wosFullResult(DOI)
}
\arguments{
\item{DOI}{a DOI}
}
\value{
result
}
\description{
Use DOI to extract metadata
}
