% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_allAuthKeywords}
\alias{get_allAuthKeywords}
\title{Extracts all author keywords from the metadata results}
\usage{
get_allAuthKeywords(wosFullResult)
}
\arguments{
\item{wosFullResult}{list, list of metadata}
}
\value{
list of keywords
}
\description{
Extracts all author keywords from the metadata results
}
