% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_citation_dataframe}
\alias{get_citation_dataframe}
\title{Read .csv files to create a citation data.frame}
\usage{
get_citation_dataframe(csv.files)
}
\arguments{
\item{csv.files}{character, file names}
}
\value{
A data.frame
}
\description{
Read .csv files to create a citation data.frame
}
