% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_countries}
\alias{get_countries}
\title{Extract the list of possible \code{country}}
\usage{
get_countries(df, country_type = "Country_current")
}
\arguments{
\item{df}{a data.frame}

\item{country_type}{character, name of the column for the different types of \code{country}}
}
\value{
character, list of unique values for the \code{country_type}
}
\description{
Extract the list of possible \code{country}
}
