% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{get_country_distance}
\alias{get_country_distance}
\title{Get the Jensen-Shannon distance across countries}
\usage{
get_country_distance(probs)
}
\arguments{
\item{probs}{probabilities}
}
\value{
data.frame
}
\description{
Get the Jensen-Shannon distance across countries
}
