% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_csv_files}
\alias{get_csv_files}
\title{List .csv files in a directory}
\usage{
get_csv_files(csv.dir)
}
\arguments{
\item{csv.dir}{character or file.path, target directory}
}
\value{
character, file names
}
\description{
List .csv files in a directory
}
