% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_mining.R
\name{get_dtm}
\alias{get_dtm}
\title{Get document-term matrix from a document-feature matrix and a list of tokens}
\usage{
get_dtm(obj_dfm, filtering_tokens)
}
\arguments{
\item{obj_dfm}{a document-feature matrix}

\item{filtering_tokens}{a list of tokens which are match against the features of the document-feature matrix}
}
\value{
a document-term matrix with nrow == nrow(obj_dfm) and ncol == length(filtering_tokens)
}
\description{
Get document-term matrix from a document-feature matrix and a list of tokens
}
