% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endnote_processing.R
\name{get_endnote_titles}
\alias{get_endnote_titles}
\title{Retrieve the titles of the documents in the English corpus}
\usage{
get_endnote_titles(language)
}
\arguments{
\item{language}{character, one of "english", "portuguese", "spanish"}
}
\value{

}
\description{
Retrieve the titles of the documents in the English corpus
}
