% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endnote_processing.R
\name{get_endnote_xml}
\alias{get_endnote_xml}
\title{Parse the .xml database from EndNote}
\usage{
get_endnote_xml(language)
}
\arguments{
\item{language}{character, one of "english", "portuguese", "spanish"}
}
\value{
xmldf, a data.frame with the data extracted by EndNote
}
\description{
Parse the .xml database from EndNote
}
