% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_EndNoteIdcorpus}
\alias{get_EndNoteIdcorpus}
\title{get document ID from EndNote query corpus database}
\usage{
get_EndNoteIdcorpus(in_corpus)
}
\arguments{
\item{in_corpus}{corpus database}
}
\value{
list of document IDs
}
\description{
get document ID from EndNote query corpus database
}
