% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_ind_hasCountryTag}
\alias{get_ind_hasCountryTag}
\title{Identifies if a document has a tag}
\usage{
get_ind_hasCountryTag(boolean_AuthKeywords)
}
\arguments{
\item{boolean_AuthKeywords}{data.frame of logical values indicating if a country has been found in the author keywords}
}
\value{
list of logical
}
\description{
Identifies if a document has a tag
}
