% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{get_JSd_corpus}
\alias{get_JSd_corpus}
\title{Calculates the Jensen-Shannon distance for countries}
\usage{
get_JSd_corpus(top, .plot = FALSE)
}
\arguments{
\item{.plot}{logical, if \code{TRUE} returns a plot, if not returns the Jensen-Shannon distance}

\item{df}{a data.frame}
}
\value{
the Jensen-Shannon distance or a \code{ggplot} object depending on \code{.plot}
}
\description{
Calculates the Jensen-Shannon distance for countries
}
