% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{get_JSd_country}
\alias{get_JSd_country}
\title{Calculates the Jensen-Shannon distance for countries}
\usage{
get_JSd_country(top, .df = df, .plot = FALSE)
}
\arguments{
\item{top}{one possible value for the \code{topic} column of the data.frame}

\item{.df}{a data.frame}

\item{.plot}{logical, if \code{TRUE} returns a plot, if not returns the Jensen-Shannon distance}
}
\value{
the Jensen-Shannon distance or a \code{ggplot} object depending on \code{.plot}
}
\description{
Calculates the Jensen-Shannon distance for countries
}
