% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_language_dfs}
\alias{get_language_dfs}
\title{Read the citation data frame and store them into a named list}
\usage{
get_language_dfs(languages)
}
\arguments{
\item{languages}{character, of one "english", "spanish", "portuguese"}
}
\value{
a named list with \code{length(languages)} elements, each a data.frame
}
\description{
Read the citation data frame and store them into a named list
}
