% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emails.R
\name{get_mail}
\alias{get_mail}
\title{Extract email addresses from \code{pdf} documents
You probably want to execute this code on a linux server to avoid the issues with special character handling on Windows}
\usage{
get_mail(pdf)
}
\arguments{
\item{pdf}{path to a \code{.pdf} files}
}
\value{
emails addresses found in the \code{.pdf} document
}
\description{
Extract email addresses from \code{pdf} documents
You probably want to execute this code on a linux server to avoid the issues with special character handling on Windows
}
