% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_meta_df}
\alias{get_meta_df}
\title{Binds the separate languages data.frame into a meta data.frame}
\usage{
get_meta_df(language_dfs)
}
\arguments{
\item{language_dfs}{a named list of data.frame}
}
\value{
a data.frame
}
\description{
Binds the separate languages data.frame into a meta data.frame
}
