% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get_MLDR}
\alias{get_MLDR}
\title{internal function to get MLDR}
\usage{
get_MLDR(trainingData, validationHumanReadingDTM)
}
\arguments{
\item{trainingData}{data.frame of training data}

\item{validationHumanReadingDTM}{document-term matrix from human reading}
}
\value{
MLDR a multilabel data.frame from mldr package
}
\description{
internal function to get MLDR
}
