% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{get_network}
\alias{get_network}
\title{Creates the adjencency matrix of a bi-partite network: country to topics}
\usage{
get_network(
  type = "theme",
  prob = TRUE,
  filter_method = FALSE,
  blindspot = FALSE,
  country.threshold = 30,
  percentile.threshold = 0.9,
  country_filter = FALSE,
  countries = NULL,
  theme_filter = FALSE,
  themes = NULL
)
}
\arguments{
\item{type}{one of "methods", "NSF_general", "NSF_specific", "spatial scale", "theme", "water budget"}

\item{prob}{logical, define if the country labels are to be treated as probabilities}

\item{filter_method}{logical, controls if methods topics are removed}

\item{blindspot}{logical, controls if the network shows blindspots, if so the network is inverted}

\item{country.threshold}{numeric, minimum number of documents per country}

\item{percentile.threshold}{numeric, controls at which percentile the adjacency matrix is filtered for visualization}

\item{country_filter}{logical, controls if countries should be filtered, see \code{countries}}

\item{countries}{character, filter for the provided countries}

\item{theme_filter}{logical, controls if themes should be filtered, see \code{themes}}

\item{themes}{character, filter for the provided themes}
}
\value{
a matrix
}
\description{
Creates the adjencency matrix of a bi-partite network: country to topics
}
