% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{get_non_duplicate_pdfs}
\alias{get_non_duplicate_pdfs}
\title{Get the indices of duplicated pdf}
\usage{
get_non_duplicate_pdfs(names)
}
\arguments{
\item{names}{character, pdf names}
}
\value{
a list of indices where names are not duplicated
}
\description{
Get the indices of duplicated pdf
}
