% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formatting.R
\name{get_pdf_files}
\alias{get_pdf_files}
\title{Get the list of files in a given directory}
\usage{
get_pdf_files(root.dir, pdf.dir)
}
\arguments{
\item{root.dir}{character or file.path pointing to a root directory}

\item{pdf.dir}{character of file.path pointing to a directory containing the pdf files, defined relative to root.dir}
}
\value{
a named list with two elements: \code{names} with the short names of the pdf files and \code{full.names} with the complete path to the pdf files.
}
\description{
Get the list of files in a given directory
}
