% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get_ps}
\alias{get_ps}
\title{Get the parameter set to tune to for a given learner}
\usage{
get_ps(lrn.id, data, grid_resolution)
}
\arguments{
\item{lrn.id}{learner id from MLR}

\item{data}{data used to construst the tuning values of the hyperparameters}

\item{grid_resolution}{resolution of the tuning grid}
}
\value{
a MLR parameter set
}
\description{
Get the parameter set to tune to for a given learner
}
