% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_relevantCountries}
\alias{get_relevantCountries}
\title{Extract countries names to be searched for in the author keywords}
\usage{
get_relevantCountries()
}
\value{
List of country names
}
\description{
Extract countries names to be searched for in the author keywords
}
