% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rootdir.R
\name{get_rootdir}
\alias{get_rootdir}
\title{Convenience functions to allow for inter-operability between different systems}
\usage{
get_rootdir()
}
\value{
a root file.path depending on the computer name
}
\description{
Convenience functions to allow for inter-operability between different systems
}
