% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_download.R
\name{get_samples}
\alias{get_samples}
\title{Select the documents for downloading while correcting for bias in terms of year and sources}
\usage{
get_samples(language, n, pl = FALSE)
}
\arguments{
\item{language}{character, one of "enlish", "portuguese", "spanish"}

\item{n}{the number of samples to select}

\item{pl}{logical, controls if a plot is produced or not}
}
\value{
samples indices
}
\description{
Select the documents for downloading while correcting for bias in terms of year and sources
}
