% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_scopusAbstract}
\alias{get_scopusAbstract}
\title{Parse htlm page from Scopus to retrieve the abstract of a document}
\usage{
get_scopusAbstract(urlScopus)
}
\arguments{
\item{urlScopus}{URL of a document}
}
\value{
Text of the abstract of the document
}
\description{
Parse htlm page from Scopus to retrieve the abstract of a document
}
