% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get_short_long_term_pred}
\alias{get_short_long_term_pred}
\title{Legacy function to assess performance of learner that learned the fine temporal scale against the aggregated temporal scale}
\usage{
get_short_long_term_pred(lrn)
}
\arguments{
\item{lrn}{learner, used to access associated benchmark predictions}
}
\description{
Legacy function to assess performance of learner that learned the fine temporal scale against the aggregated temporal scale
}
