% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_titleDocs}
\alias{get_titleDocs}
\title{Read topic model file titles}
\usage{
get_titleDocs(
  topicDocs,
  fname = system.file("extdata", "info.dat", package = "wateReview")
)
}
\arguments{
\item{fname}{path to the topic model file titles}
}
\value{
topic model file titles
}
\description{
Read topic model file titles
}
