% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_titleInd}
\alias{get_titleInd}
\title{This function performs the cross-walk between human reading databases and topic model databases using document titles}
\usage{
get_titleInd(
  titleInd_file = system.file("extdata", "titleInd.Rds", package = "wateReview"),
  humanReadingDatabase = NULL,
  topicModelTitles = NULL
)
}
\arguments{
\item{titleInd_file}{a filename, if it exists, file content is read, if it does not exist the formatting and regexpr matching is performed}

\item{humanReadingDatabase}{human reading database with column "title", default to NULL}

\item{topicModelTitles}{titles of the topic model files, default to NULL}
}
\value{
matching indices between the two databases
}
\description{
This function performs the cross-walk between human reading databases and topic model databases using document titles
}
