% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_colors.R
\name{get_topColors}
\alias{get_topColors}
\title{Get the dominant colors from country flags
The country flag are extracted from http://hjnilsson.github.io/country-flags/}
\usage{
get_topColors(iso, n = 3)
}
\arguments{
\item{iso}{ISO Alpha-2 code}

\item{n}{number of colors}
}
\value{
top \code{n} colors
}
\description{
Get the dominant colors from country flags
The country flag are extracted from http://hjnilsson.github.io/country-flags/
}
