% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{get_topic_distance}
\alias{get_topic_distance}
\title{Get the Jensen-Shannon distance across topics}
\usage{
get_topic_distance(probs)
}
\arguments{
\item{probs}{probabilities}
}
\value{
data.frame
}
\description{
Get the Jensen-Shannon distance across topics
}
