% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_validationHumanReading}
\alias{get_validationHumanReading}
\title{Read human reading database}
\usage{
get_validationHumanReading(scale_type = "location")
}
\arguments{
\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}
}
\value{
validation human-read data
}
\description{
Read human reading database
}
