% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_webscrapped_trainingLabels}
\alias{get_webscrapped_trainingLabels}
\title{read webscrapped training labels}
\usage{
get_webscrapped_trainingLabels(fname = "webscrapped_trainingLabels.Rds")
}
\arguments{
\item{fname}{path to webscrapped training labels data}
}
\value{
webscrapped training labels
}
\description{
read webscrapped training labels
}
