% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_webscrapped_validationDTM}
\alias{get_webscrapped_validationDTM}
\title{read training data (document-term matrix corresponding to webscrapped labels)}
\usage{
get_webscrapped_validationDTM(
  fname = system.file("extdata", "webscrapped_validationDTM.Rds", package =
    "wateReview")
)
}
\arguments{
\item{fname}{path to training data (document-term matrix corresponding to webscrapped labels)}
}
\value{
training data (document-term matrix corresponding to webscrapped labels)
}
\description{
read training data (document-term matrix corresponding to webscrapped labels)
}
