% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_wosAbstract}
\alias{get_wosAbstract}
\title{Use Elsevier API to retrieve the abstract of a document}
\usage{
get_wosAbstract(DOI)
}
\arguments{
\item{DOI}{DOI (e.g. from Web of Science database)}
}
\value{
Text of the abstract of the document
}
\description{
Use Elsevier API to retrieve the abstract of a document
}
