% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{get_wosAuthKeywords}
\alias{get_wosAuthKeywords}
\title{Use Elsevier API to retrieve the author keywords of a document}
\usage{
get_wosAuthKeywords(DOI, from_FullResult = FALSE)
}
\arguments{
\item{DOI}{DOI (e.g. from Web of Science database) or the complete result from a previous request}

\item{from_Fullresult}{logical, if TRUE extracts the abstract from the result of a previous request}
}
\value{
List of author keywords of the corresponding document
}
\description{
Use Elsevier API to retrieve the author keywords of a document
}
