% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{get_wrappedLearnersList}
\alias{get_wrappedLearnersList}
\title{Create a list of wrapped learners}
\usage{
get_wrappedLearnersList(learnerList, data, mes, grid_resolution)
}
\arguments{
\item{learnerList}{a list of learners}

\item{data}{used to construct the value of hyperparameters}

\item{mes}{measures}

\item{grid_resolution}{resolution of the tuning grid}
}
\value{
list of wrapped learners
}
\description{
Create a list of wrapped learners
}
