% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{join_database_shapefile}
\alias{join_database_shapefile}
\title{Assign the data from the country database to the country shapefile}
\usage{
join_database_shapefile(countries_database, countries_shapefile)
}
\arguments{
\item{countries_database}{a data.frame with the country database and with a \code{COUNTRY} column}

\item{countries_shapefile}{a raster::shapefile file with a \code{COUNTRY} attribute}
}
\value{
a SpatialPolygonsDataFrame with the country database adjoined
}
\description{
Assign the data from the country database to the country shapefile
}
