% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{make.countryNetwork}
\alias{make.countryNetwork}
\title{Creates a weighted adjacency matrix of probability of citation between countries}
\usage{
make.countryNetwork(source_ids, citation_network)
}
\arguments{
\item{source_ids}{source_ids}

\item{citation_network}{an edge data.frame}
}
\value{
a weighted adjacency matrix of probability of citation between countries
}
\description{
Creates a weighted adjacency matrix of probability of citation between countries
}
