% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make.targetData}
\alias{make.targetData}
\title{Create the target data to predict from}
\usage{
make.targetData(DTM, addTopicDocs = FALSE, topicDocs = NULL)
}
\arguments{
\item{DTM}{the complete document-term matrix}
}
\value{
a data.frame
}
\description{
Create the target data to predict from
}
