% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make.task}
\alias{make.task}
\title{Make an MLR task}
\usage{
make.task(trainingData, validationHumanReadingDTM, model_type, filter = FALSE)
}
\arguments{
\item{trainingData}{training data}

\item{validationHumanReadingDTM}{training data from human reading, used to extract target column names}

\item{model_type}{type of predictions}

\item{filter}{logical, if true create task for a binary classification Irrelevant/Relevant}
}
\value{
a MLR task
}
\description{
Make an MLR task
}
