% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make.trainingDataMulticlass}
\alias{make.trainingDataMulticlass}
\title{Create training data for a multiclass classification}
\usage{
make.trainingDataMulticlass(
  trainingData,
  validationHumanReadingDTM,
  humanReadingTrainingLabels,
  webscrapped_validationDTM,
  webscrapped_trainingLabels,
  filter = FALSE,
  addWebscrapped = FALSE,
  obs_threshold = 10,
  filterIrrelevant = TRUE,
  addTopicDocs = FALSE,
  validationTopicDocs = NULL
)
}
\arguments{
\item{trainingData}{the already aggreated validation data}

\item{validationHumanReadingDTM}{document-term matrix from human reading}

\item{humanReadingTrainingLabels}{labels from human-reading}

\item{webscrapped_validationDTM}{document-term matrix from webscrapping}

\item{webscrapped_trainingLabels}{labels from webscrapping}

\item{filter}{logical, if true create training data for a binary classification Irrelevant/Relevant}
}
\value{
a data.frame containing the training data with a target column "countryLabelFilter" or "countryLabel"
}
\description{
Create training data for a multiclass classification
}
