% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make_AUCPlot}
\alias{make_AUCPlot}
\title{Create a comparison plot of AUC}
\usage{
make_AUCPlot(bmr, binary = FALSE)
}
\arguments{
\item{bmr}{benchmark results from MLR}

\item{binary}{logical, if TRUE it's a binary classification and the measure name has to be changed}
}
\value{
named list: p_auc: the plot; medians: a table with median value of auc
}
\description{
Create a comparison plot of AUC
}
