% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_corrmatrix}
\alias{make_corrmatrix}
\title{Make a correlation matrix}
\usage{
make_corrmatrix(
  data_type = "country_descriptors",
  corr_type = "spearman",
  rescale = FALSE,
  .plot = TRUE
)
}
\arguments{
\item{data_type}{character, of one "country_descriptors", "survey_results", "survey_results_pct", "all"; default to "country_descriptors"}

\item{corr_type}{character, one of "spearman", "pearson"}

\item{rescale}{logical, default to \code{FALSE}, controls if the data have to be rescaled}

\item{.plot}{logical, controls if the correlation plot is outputed}
}
\value{
a correlation matrix
}
\description{
Make a correlation matrix
}
