% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{make_country_tokens}
\alias{make_country_tokens}
\title{Tokenize labels, here a list of relevant countries}
\usage{
make_country_tokens(webscrapped_trainingLabels)
}
\arguments{
\item{webscrapped_trainingLabels}{data.frame of multiple labels}
}
\value{
Tokenized labels
}
\description{
Tokenize labels, here a list of relevant countries
}
