% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_dendrogram}
\alias{make_dendrogram}
\title{Make a dendrogram}
\usage{
make_dendrogram(rescaled_data, num_clusters = 3, check_na = FALSE)
}
\arguments{
\item{rescaled_data}{a data.frame with data}

\item{num_clusters}{numeric, number of clusters}

\item{check_na}{logical, default to \code{FALSE}, controls if the data have to be checked for \code{NA}}
}
\value{
a dendrogram
}
\description{
Make a dendrogram
}
