% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make_df_docs}
\alias{make_df_docs}
\title{Creates a subset of the LDA topics dataframe into theme dataframe}
\usage{
make_df_docs(
  theme_type = "theme",
  description = NULL,
  save = FALSE,
  topicDocs = .topicDocs
)
}
\arguments{
\item{theme_type}{one of topic_name, theme, NSF_specific, NSF_general, passed to make_df_docs}

\item{description}{one of spatial_scale, methods, water budget, passed to make_df_docs}

\item{save}{boolean, if true, write the results, if false, return the theme dataframe}

\item{topicDocs}{the output from the topicdata, aligned with the corpus}
}
\description{
Creates a subset of the LDA topics dataframe into theme dataframe
}
