% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_hist}
\alias{make_hist}
\title{Makes a density plot of a selected attribute}
\usage{
make_hist(attribute_name, shp = countries_shapefile)
}
\arguments{
\item{attribute_name}{character, name of an attribute to create a density plot}

\item{shp}{a SpatialPolygonsDataFrame with attributes}
}
\value{
a \code{ggplot2} object
}
\description{
Makes a density plot of a selected attribute
}
