% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make_humanReadingTrainingLabels}
\alias{make_humanReadingTrainingLabels}
\title{Create training labels from aligned human reading database}
\usage{
make_humanReadingTrainingLabels(
  validationHumanReading,
  scale_type = "location",
  webscrapped_trainingLabels
)
}
\arguments{
\item{validationHumanReading}{aligned human reading database}

\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}

\item{webscrapped_trainingLabels}{webscrapped training labels used for comparison purpose only}
}
\value{
training labels
}
\description{
Create training labels from aligned human reading database
}
