% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{make_map}
\alias{make_map}
\title{Makes a map for a give attribute}
\usage{
make_map(attribute, title = NULL)
}
\arguments{
\item{attribute}{character, name of an attribute to map}

\item{title}{optional, default to \code{NULL}, costum title for the map}
}
\value{
a tmap object
}
\description{
Makes a map for a give attribute
}
