% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Make randomForest prediction}
\usage{
make_predictions(
  lrn,
  parvals,
  trainingData,
  targetData,
  model_type,
  filter = TRUE
)
}
\arguments{
\item{lrn}{a learner name}

\item{parvals}{a list of best tuned hyperparameters}

\item{trainingData}{data to train the model}

\item{targetData}{data used to predict}

\item{model_type}{}

\item{filter}{logical, if true will train and predict for the relevance filter}
}
\value{
prediction or membership
}
\description{
Make randomForest prediction
}
