% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endnote_processing.R
\name{make_pretty_str}
\alias{make_pretty_str}
\title{This function takes care of some formatting issue that appeared in the process of aligning database coming from the query and from EndNote.
The function removes alpha-numeric characters, some special characters, trim whitespaces and concatenate them.}
\usage{
make_pretty_str(string_list)
}
\arguments{
\item{string_list}{a list of string}
}
\value{
a cleaned list of string
}
\description{
This function takes care of some formatting issue that appeared in the process of aligning database coming from the query and from EndNote.
The function removes alpha-numeric characters, some special characters, trim whitespaces and concatenate them.
}
