% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{make_topicNetwork}
\alias{make_topicNetwork}
\title{Creates a weighted adjacency matrix of probability of citation between topics}
\usage{
make_topicNetwork(
  source_ids,
  citation_network,
  type = "NSF_specific",
  percentile.threshold = 0.9
)
}
\arguments{
\item{source_ids}{source_ids}

\item{citation_network}{an edge data.frame}
}
\value{
a weighted adjacency matrix of probability of citation between topics
}
\description{
Creates a weighted adjacency matrix of probability of citation between topics
}
