% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{make_trainingData}
\alias{make_trainingData}
\title{Create training data for a multilabel classification}
\usage{
make_trainingData(
  validationHumanReadingDTM,
  humanReadingTrainingLabels,
  webscrapped_validationDTM,
  webscrapped_trainingLabels,
  scale_type = "location",
  aggregate_labels = FALSE
)
}
\arguments{
\item{validationHumanReadingDTM}{document-term matrix from human reading}

\item{humanReadingTrainingLabels}{labels from human-reading}

\item{webscrapped_validationDTM}{document-term matrix from webscrapping}

\item{webscrapped_trainingLabels}{labels from webscrapping}

\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}

\item{aggregate_labels}{logical, for temporal scale, option to aggregate into three larger classes}
}
\value{
a data.frame with nrow == nrow(validationHumanReading) + nrow(webscrapped_validationDTM) and ncol == ncol(validationHumanReadingDTM) + ncol(humanReadingTrainingLabels)
}
\description{
Create training data for a multilabel classification
}
