% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{make_webscrapped_trainingData}
\alias{make_webscrapped_trainingData}
\title{Create a data.frame of multilabels using the webscrapped author keywords}
\usage{
make_webscrapped_trainingData(
  boolean_AuthKeywords,
  ind_hasCountryTag,
  englishCorpus,
  englishCorpus_file
)
}
\arguments{
\item{boolean_AuthKeywords}{data.frame of multilabels webscrapped author keywords}

\item{ind_hasCountryTag}{list of boolean indicating if an entry has at least one label}

\item{englishCorpus}{databse of corpus of document with abstracts}

\item{englishCorpus_file}{file path to the complete corpus}
}
\value{
list with 3 elements: country_tokens: tokenized country labels, webscrapped_validationDTM: a document term matrix derived from the tokenized country labels, webscrapped_trainingLabels: webscrapped multilabels
}
\description{
Create a data.frame of multilabels using the webscrapped author keywords
}
