% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{melt_df_country}
\alias{melt_df_country}
\title{Manually melt a data.frame}
\usage{
melt_df_country(df_country, country_type = "Country_current")
}
\arguments{
\item{df_country}{a country data.frame}

\item{country_type}{character, name of the column for the different types of \code{country}}
}
\value{
a data.frame
}
\description{
Manually melt a data.frame
}
