% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{multiclassBenchmark}
\alias{multiclassBenchmark}
\title{Perform a benchmark between non-tuned algorithm adaptation and multilabel and binary relevance wrappers}
\usage{
multiclassBenchmark(trainingData, model_type, filter = FALSE, tune = NULL)
}
\arguments{
\item{trainingData}{training data}

\item{model_type}{type of predictions}

\item{filter}{logical, if true create task for a binary classification Irrelevant/Relevant}
}
\value{
benchmark object
}
\description{
Perform a benchmark between non-tuned algorithm adaptation and multilabel and binary relevance wrappers
}
