% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{multilabelBenchmark}
\alias{multilabelBenchmark}
\title{Perform a benchmark between algorithm adaptation and multilabel and binary relevance wrappers}
\usage{
multilabelBenchmark(
  trainingData,
  validationHumanReadingDTM,
  model_type,
  scale_type = "location",
  aggregated_labels = FALSE,
  obs_threshold = 10
)
}
\arguments{
\item{trainingData}{training data}

\item{validationHumanReadingDTM}{training data from human reading, used to extract target column names}

\item{model_type}{type of predictions}

\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}

\item{aggregated_labels}{logical}

\item{obs_threshold}{remove columns with less than this threshold}
}
\value{
benchmark object
}
\description{
Perform a benchmark between algorithm adaptation and multilabel and binary relevance wrappers
}
