% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{normalize_adj_matrix}
\alias{normalize_adj_matrix}
\title{Normalize an adjacency matrix by its rows (i.e., "from")}
\usage{
normalize_adj_matrix(adj)
}
\arguments{
\item{adj}{a matrix}
}
\value{
a normalized matrix
}
\description{
Normalize an adjacency matrix by its rows (i.e., "from")
}
