% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{order.data}
\alias{order.data}
\title{Order data to LDA order}
\usage{
order.data(data, EndNoteIdLDA, EndNoteIdcorpus)
}
\arguments{
\item{data}{data: corpus of documents, indices of tagged document, ...}

\item{EndNoteIdLDA}{document id in the corpus from the query}

\item{EndNoteIdcorpus}{document id in the topic model}
}
\value{
data with matching records in both databases and webscrapped abstracts
}
\description{
Order data to LDA order
}
