% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{PerfVisMultilabel}
\alias{PerfVisMultilabel}
\title{Make a performance plot for multilabel classification}
\usage{
PerfVisMultilabel(AggrPerformances)
}
\arguments{
\item{AggrPerformances}{data.frame of aggregated performance from getBMRAggrPerformances}
}
\value{
ggplot plot
}
\description{
Make a performance plot for multilabel classification
}
