% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{plot_df_country}
\alias{plot_df_country}
\title{Plot each variables in a melted data.frame in a faceted \code{ggplot}}
\usage{
plot_df_country(df_country, title_text = "Brazil")
}
\arguments{
\item{df_country}{a melted data.frame with a \code{variable} column}

\item{title_text}{character, to be used as title for the graph}
}
\description{
Plot each variables in a melted data.frame in a faceted \code{ggplot}
}
