% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_download.R
\name{print_estimate}
\alias{print_estimate}
\title{Print the time estimate for manual downloading}
\usage{
print_estimate(n_dl = 8, time = 5, pr = FALSE)
}
\arguments{
\item{n_dl}{numeric, number of downloaders}

\item{time}{numeric, time to download one document in minutes}

\item{pr}{logical, controls the output of the values of the estimated times}
}
\value{
The time estimates in hours
}
\description{
Print the time estimate for manual downloading
}
