% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{QA.alignedData}
\alias{QA.alignedData}
\title{Perform QA/QC on aligned data}
\usage{
QA.alignedData(alignedData, scale_type = "location")
}
\arguments{
\item{alignedData}{list of aligned data between human reading and topic model}

\item{scale_type}{One of "location", "spatial", "temporal", default to "location"}
}
\value{
list with three elements: validationHumanReading, validationTopicDocs, validationDTM
}
\description{
Perform QA/QC on aligned data
}
