% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{QA.AuthKeywords}
\alias{QA.AuthKeywords}
\title{Quality analysis on the author keywords retrieved}
\usage{
QA.AuthKeywords(wosAuthKeywords, relevant_countries)
}
\arguments{
\item{wosAuthKeywords}{list, list of author keywords}

\item{relevant_countries}{list, list of country names}
}
\value{
list of failed entries
}
\description{
Quality analysis on the author keywords retrieved
}
