% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{QA.EndNoteIdCorpusLDA}
\alias{QA.EndNoteIdCorpusLDA}
\title{Performs a quick QA on both sources for document ID}
\usage{
QA.EndNoteIdCorpusLDA(EndNoteIdLDA, EndNoteIdcorpus)
}
\value{
none
}
\description{
Performs a quick QA on both sources for document ID
}
