% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_predictions.R
\name{QA.oldXnewPredictions}
\alias{QA.oldXnewPredictions}
\title{this function performs a quick quality analysis and difference with historical predictions}
\usage{
QA.oldXnewPredictions(predCountry)
}
\arguments{
\item{predCountry}{the predicted country locations}
}
\description{
this function performs a quick quality analysis and difference with historical predictions
}
