% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{QA_EndNoteIdCorpusLDA}
\alias{QA_EndNoteIdCorpusLDA}
\title{Performs a quick QA on both sources for document ID}
\usage{
QA_EndNoteIdCorpusLDA(EndNoteIdLDA, EndNoteIdcorpus)
}
\arguments{
\item{EndNoteIdLDA}{document id in the corpus from the query}

\item{EndNoteIdcorpus}{document id in the topic model}
}
\value{
none
}
\description{
Performs a quick QA on both sources for document ID
}
