% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_QA.R
\name{query_QA_plots}
\alias{query_QA_plots}
\title{Make QA/QC plot for a given data.frame by comparing corpus present in the query and actually collected}
\usage{
query_QA_plots(meta_df, languages = c("english", "portuguese", "spanish"))
}
\arguments{
\item{meta_df}{a data.frame with information about both the query and the collected corpus}

\item{languages}{character, vector of possible languages}
}
\value{
a named list with 5 plots: \code{source}, \code{source_DOI}, \code{time_histogram}, \code{time_density}, \code{citations}
}
\description{
Make QA/QC plot for a given data.frame by comparing corpus present in the query and actually collected
}
