% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_QA.R
\name{read_citation_dataframe}
\alias{read_citation_dataframe}
\title{Read the citation data.frame exported by the query processing}
\usage{
read_citation_dataframe(language)
}
\arguments{
\item{language}{character, one of "english", "spanish", "portuguese"}
}
\value{
data.frame
}
\description{
Read the citation data.frame exported by the query processing
}
