% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{read_countries_database}
\alias{read_countries_database}
\title{Read the country database file}
\usage{
read_countries_database(file = "countries_database_final.csv")
}
\arguments{
\item{file}{character, a filename to fetch into \code{extdata}}
}
\value{
a data.frame
}
\description{
Read the country database file
}
