% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{read_countries_shapefile}
\alias{read_countries_shapefile}
\title{Read the country shapefile}
\usage{
read_countries_shapefile(file = "LAC.shp")
}
\arguments{
\item{file}{character, a filename to fetch into \code{extdata}}
}
\value{
a SpatialPolygonsDataFrame
}
\description{
Read the country shapefile
}
