% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{read_survey_df}
\alias{read_survey_df}
\title{Read the survey data}
\usage{
read_survey_df(fname = "Water Management in Latin America_edit.csv")
}
\arguments{
\item{fname}{file.path to the survey data, expecting a \code{.csv} file}
}
\value{
a data.frame
}
\description{
Read the survey data
}
