% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{reduce_docs_for_JSd}
\alias{reduce_docs_for_JSd}
\title{Reduce the documents before calculating the Jensen-Shannon distance}
\usage{
reduce_docs_for_JSd(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame with columns \code{country}, \code{topic} and \code{value}
}
\description{
Reduce the documents before calculating the Jensen-Shannon distance
}
