% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{remove_country}
\alias{remove_country}
\title{Remove \code{country} column}
\usage{
remove_country(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame without the \code{country} column
}
\description{
Remove \code{country} column
}
