% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{remove_irrelevant}
\alias{remove_irrelevant}
\title{Removes \code{Irrelevant} entries from the \code{country} column of a data.frame}
\usage{
remove_irrelevant(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame without \code{country} entries marked as \code{Irrelevant}
}
\description{
Removes \code{Irrelevant} entries from the \code{country} column of a data.frame
}
