% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{remove_year}
\alias{remove_year}
\title{Remove \code{year} column}
\usage{
remove_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame without the \code{year} column
}
\description{
Remove \code{year} column
}
