% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{remove_year_country}
\alias{remove_year_country}
\title{Remove \code{year} and \code{country} columns}
\usage{
remove_year_country(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a data.frame without the \code{year} and \code{country} columns
}
\description{
Remove \code{year} and \code{country} columns
}
