% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_colors.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert rgb to hex colors}
\usage{
rgb2hex(r, g, b)
}
\arguments{
\item{r}{red value}

\item{g}{green value}

\item{b}{blue value}
}
\value{
hex code
}
\description{
Convert rgb to hex colors
}
