% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{select_data}
\alias{select_data}
\title{Select the type of data to display}
\usage{
select_data(data_type = "country_descriptors")
}
\arguments{
\item{data_type}{character, of one "country_descriptors", "survey_results", "survey_results_pct", "all"; default to "country_descriptors"}
}
\value{
a matrix
}
\description{
Select the type of data to display
}
