% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{select_list}
\alias{select_list}
\title{Convenience function for the shinyApp}
\usage{
select_list(data_type)
}
\arguments{
\item{data_type}{a character}
}
\value{
numeric index
}
\description{
Convenience function for the shinyApp
}
