% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_QA.R
\name{source_plot}
\alias{source_plot}
\title{Produce a barplot of the different sources present in the query and in the collected corpus}
\usage{
source_plot(meta_df, language, n = 40, facet = TRUE)
}
\arguments{
\item{meta_df}{data.frame with data of multiple \code{language}}

\item{language}{character, one of "english", "spanish", "portuguese"}

\item{n}{numeric, number of bars to display}

\item{facet}{logical, controls if the plot is faceted by DOI presence}
}
\value{
\code{ggplot} object
}
\description{
Produce a barplot of the different sources present in the query and in the collected corpus
}
