% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{transform_data}
\alias{transform_data}
\title{Transform the data with centering, scaling and Box-Cox transformations}
\usage{
transform_data(data, methods = c("center", "scale", "BoxCox"))
}
\arguments{
\item{data}{a data.frame}

\item{methods}{character, methods for the \code{caret::preProcess} functions}
}
\value{
a data.frame with transformed data
}
\description{
Transform the data with centering, scaling and Box-Cox transformations
}
