% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_processing.R
\name{write_citation_dataframe}
\alias{write_citation_dataframe}
\title{Save a data.frame.}
\usage{
write_citation_dataframe(csv.dir)
}
\arguments{
\item{csv.dir}{character or file.path to the location where to save the citation dataframe}
}
\description{
Save a data.frame.
}
