% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{align.dataWithEndNoteIdcorpus}
\alias{align.dataWithEndNoteIdcorpus}
\title{Align databases based on shared ID}
\usage{
align.dataWithEndNoteIdcorpus(data, EndNoteIdcorpus, EndNoteIdLDA)
}
\arguments{
\item{EndNoteIdLDA}{}
}
\value{
EndNoteIdcorpus with matching records in both databases
}
\description{
Align databases based on shared ID
}
