% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webscrapping.R
\name{align.dataWithEndNoteIdLDA}
\alias{align.dataWithEndNoteIdLDA}
\title{Align databases based on shared ID}
\usage{
align.dataWithEndNoteIdLDA(data, EndNoteIdLDA, EndNoteIdcorpus)
}
\arguments{
\item{EndNoteIdLDA}{document id in the corpus from the query}

\item{EndNoteIdcorpus}{document id in the topic model}
}
\value{
EndNoteIdLDA with matching records in both databases
}
\description{
Align databases based on shared ID
}
