% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_DTM}
\alias{get_DTM}
\title{read document-term matrix created by the text mining code}
\usage{
get_DTM(
 
    dtm_file = "F:/hguillon/research/exploitation/R/latin_america/data/obj_dtm_from_dfm_country.Rds"
)
}
\arguments{
\item{dtm_file}{path to saved document-term matrix}
}
\value{
document-term matrix
}
\description{
read document-term matrix created by the text mining code
}
