% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_formatting.R
\name{get_non_duplicate_pdfs}
\alias{get_non_duplicate_pdfs}
\title{Get the indices of duplicated documents}
\usage{
get_non_duplicate_pdfs(names)
}
\arguments{
\item{names}{character, document names}
}
\value{
a list of indices of non-duplicate documents
}
\description{
Get the indices of duplicated documents
}
