% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_formatting.R
\name{get_pdf_files}
\alias{get_pdf_files}
\title{Get the list of files in a given directory}
\usage{
get_pdf_files(pdf_dir)
}
\arguments{
\item{pdf_dir}{character of file.path pointing to a directory containing the document pdf files}
}
\value{
a named list with two elements: \code{names} with the short names of the pdf files and \code{full_names} with the complete path to the pdf files.
}
\description{
Get the list of files in a given directory
}
