% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{get_topicDocs}
\alias{get_topicDocs}
\title{Read topic model data}
\usage{
get_topicDocs(
  fname = system.file("extdata", "topicDocs.Rds", package = "wateReview")
)
}
\arguments{
\item{fname}{path to the topic model data}
}
\value{
topic model data
}
\description{
Read topic model data
}
